+++ qa/rolling-upgrade/src/javaRestTest/java/org/elasticsearch/upgrades/SourceModeRollingUpgradeIT.java
@@ -18,7 +18,6 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Map;
 
-import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 
 public class SourceModeRollingUpgradeIT extends AbstractRollingUpgradeTestCase {
@@ -83,20 +82,10 @@ public class SourceModeRollingUpgradeIT extends AbstractRollingUpgradeTestCase {
     private void assertDeprecationWarningForTemplate(String templateName) throws IOException {
         var request = new Request("GET", "/_migration/deprecations");
         var response = entityAsMap(client().performRequest(request));
-        if (response.containsKey("templates")) {
-            // Check the newer version of the deprecation API that contains the templates section
-            Map<?, ?> issuesByTemplate = (Map<?, ?>) response.get("templates");
-            assertThat(issuesByTemplate.containsKey(templateName), equalTo(true));
-            var templateIssues = (List<?>) issuesByTemplate.get(templateName);
-            assertThat(((Map<?, ?>) templateIssues.getFirst()).get("message"), equalTo(SourceFieldMapper.DEPRECATION_WARNING));
-        } else {
-            // Bwc version with 8.18 until https://github.com/elastic/elasticsearch/pull/120505/ gets backported, clean up after backport
-            var nodeSettings = (Map<?, ?>) ((List<?>) response.get("node_settings")).getFirst();
-            assertThat(nodeSettings.get("message"), equalTo(SourceFieldMapper.DEPRECATION_WARNING));
-            assertThat(
-                (String) nodeSettings.get("details"),
-                containsString(SourceFieldMapper.DEPRECATION_WARNING + " Affected component templates: [" + templateName + "]")
-            );
-        }
+        assertThat(response.containsKey("templates"), equalTo(true));
+        Map<?, ?> issuesByTemplate = (Map<?, ?>) response.get("templates");
+        assertThat(issuesByTemplate.containsKey(templateName), equalTo(true));
+        var templateIssues = (List<?>) issuesByTemplate.get(templateName);
+        assertThat(((Map<?, ?>) templateIssues.getFirst()).get("message"), equalTo(SourceFieldMapper.DEPRECATION_WARNING));
     }
 }
