+++ x-pack/plugin/deprecation/src/main/java/module-info.java
@@ -13,6 +13,7 @@ module org.elasticsearch.deprecation {
     requires org.apache.logging.log4j;
     requires org.apache.logging.log4j.core;
     requires log4j2.ecs.layout;
+    requires org.apache.lucene.core;
 
     exports org.elasticsearch.xpack.deprecation to org.elasticsearch.server;
     exports org.elasticsearch.xpack.deprecation.logging to org.elasticsearch.server;
