+++ x-pack/plugin/deprecation/src/main/java/org/elasticsearch/xpack/deprecation/IndexDeprecationChecker.java
@@ -10,6 +10,7 @@ import org.elasticsearch.cluster.ClusterState;
 import org.elasticsearch.cluster.metadata.IndexMetadata;
 import org.elasticsearch.cluster.metadata.IndexNameExpressionResolver;
 import org.elasticsearch.cluster.metadata.MappingMetadata;
+import org.elasticsearch.common.TriFunction;
 import org.elasticsearch.common.time.DateFormatter;
 import org.elasticsearch.common.time.LegacyFormatNames;
 import org.elasticsearch.index.IndexModule;
@@ -17,17 +18,19 @@ import org.elasticsearch.index.IndexSettings;
 import org.elasticsearch.index.IndexVersion;
 import org.elasticsearch.xpack.core.deprecation.DeprecatedIndexPredicate;
 import org.elasticsearch.xpack.core.deprecation.DeprecationIssue;
+import org.elasticsearch.xpack.core.transform.transforms.TransformConfig;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
+import java.util.Objects;
 import java.util.function.BiConsumer;
 import java.util.function.BiFunction;
 import java.util.function.Function;
+import java.util.stream.Collectors;
 
-import static org.elasticsearch.xpack.deprecation.DeprecationInfoAction.filterChecks;
 import static org.elasticsearch.xpack.deprecation.LegacyTiersDetection.DEPRECATION_COMMON_DETAIL;
 import static org.elasticsearch.xpack.deprecation.LegacyTiersDetection.DEPRECATION_HELP_URL;
 
@@ -39,20 +42,35 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
     public static final String NAME = "index_settings";
 
     private final IndexNameExpressionResolver indexNameExpressionResolver;
-    private final Map<String, List<String>> indexToTransformIds;
+    private final List<TriFunction<IndexMetadata, ClusterState, Map<String, List<String>>, DeprecationIssue>> checks = List.of(
+        this::oldIndicesCheck,
+        this::ignoredOldIndicesCheck,
+        this::translogRetentionSettingCheck,
+        this::checkIndexDataPath,
+        this::storeTypeSettingCheck,
+        this::deprecatedCamelCasePattern,
+        this::legacyRoutingSettingCheck
+    );
 
-    public IndexDeprecationChecker(IndexNameExpressionResolver indexNameExpressionResolver, Map<String, List<String>> indexToTransformIds) {
+    public IndexDeprecationChecker(IndexNameExpressionResolver indexNameExpressionResolver) {
         this.indexNameExpressionResolver = indexNameExpressionResolver;
-        this.indexToTransformIds = indexToTransformIds;
     }
 
     @Override
-    public Map<String, List<DeprecationIssue>> check(ClusterState clusterState, DeprecationInfoAction.Request request) {
+    public Map<String, List<DeprecationIssue>> check(
+        ClusterState clusterState,
+        DeprecationInfoAction.Request request,
+        TransportDeprecationInfoAction.PrecomputedData precomputedData
+    ) {
         Map<String, List<DeprecationIssue>> indexSettingsIssues = new HashMap<>();
         String[] concreteIndexNames = indexNameExpressionResolver.concreteIndexNames(clusterState, request);
+        Map<String, List<String>> indexToTransformIds = indexToTransformIds(precomputedData.transformConfigs());
         for (String concreteIndex : concreteIndexNames) {
             IndexMetadata indexMetadata = clusterState.getMetadata().index(concreteIndex);
-            List<DeprecationIssue> singleIndexIssues = filterChecks(indexSettingsChecks(), c -> c.apply(indexMetadata, clusterState));
+            List<DeprecationIssue> singleIndexIssues = checks.stream()
+                .map(c -> c.apply(indexMetadata, clusterState, indexToTransformIds))
+                .filter(Objects::nonNull)
+                .toList();
             if (singleIndexIssues.isEmpty() == false) {
                 indexSettingsIssues.put(concreteIndex, singleIndexIssues);
             }
@@ -63,24 +81,16 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
         return indexSettingsIssues;
     }
 
-    private List<BiFunction<IndexMetadata, ClusterState, DeprecationIssue>> indexSettingsChecks() {
-        return List.of(
-            this::oldIndicesCheck,
-            this::ignoredOldIndicesCheck,
-            IndexDeprecationChecker::translogRetentionSettingCheck,
-            IndexDeprecationChecker::checkIndexDataPath,
-            IndexDeprecationChecker::storeTypeSettingCheck,
-            IndexDeprecationChecker::deprecatedCamelCasePattern,
-            IndexDeprecationChecker::legacyRoutingSettingCheck
-        );
-    }
-
     @Override
     public String getName() {
         return NAME;
     }
 
-    private DeprecationIssue oldIndicesCheck(IndexMetadata indexMetadata, ClusterState clusterState) {
+    private DeprecationIssue oldIndicesCheck(
+        IndexMetadata indexMetadata,
+        ClusterState clusterState,
+        Map<String, List<String>> indexToTransformIds
+    ) {
         // TODO: this check needs to be revised. It's trivially true right now.
         IndexVersion currentCompatibilityVersion = indexMetadata.getCompatibilityVersion();
         // We intentionally exclude indices that are in data streams because they will be picked up by DataStreamDeprecationChecks
@@ -91,13 +101,13 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
                 "https://www.elastic.co/guide/en/elasticsearch/reference/master/breaking-changes-9.0.html",
                 "This index has version: " + currentCompatibilityVersion.toReleaseVersion(),
                 false,
-                meta(indexMetadata)
+                meta(indexMetadata, indexToTransformIds)
             );
         }
         return null;
     }
 
-    private Map<String, Object> meta(IndexMetadata indexMetadata) {
+    private Map<String, Object> meta(IndexMetadata indexMetadata, Map<String, List<String>> indexToTransformIds) {
         var transforms = indexToTransformIds.getOrDefault(indexMetadata.getIndex().getName(), List.of());
         if (transforms.isEmpty()) {
             return Map.of("reindex_required", true);
@@ -106,7 +116,11 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
         }
     }
 
-    private DeprecationIssue ignoredOldIndicesCheck(IndexMetadata indexMetadata, ClusterState clusterState) {
+    private DeprecationIssue ignoredOldIndicesCheck(
+        IndexMetadata indexMetadata,
+        ClusterState clusterState,
+        Map<String, List<String>> indexToTransformIds
+    ) {
         IndexVersion currentCompatibilityVersion = indexMetadata.getCompatibilityVersion();
         // We intentionally exclude indices that are in data streams because they will be picked up by DataStreamDeprecationChecks
         if (DeprecatedIndexPredicate.reindexRequired(indexMetadata, true) && isNotDataStreamIndex(indexMetadata, clusterState)) {
@@ -118,17 +132,21 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
                     + currentCompatibilityVersion.toReleaseVersion()
                     + " and will be supported as read-only in 9.0",
                 false,
-                meta(indexMetadata)
+                meta(indexMetadata, indexToTransformIds)
             );
         }
         return null;
     }
 
-    private static boolean isNotDataStreamIndex(IndexMetadata indexMetadata, ClusterState clusterState) {
+    private boolean isNotDataStreamIndex(IndexMetadata indexMetadata, ClusterState clusterState) {
         return clusterState.metadata().findDataStreams(indexMetadata.getIndex().getName()).isEmpty();
     }
 
-    private static DeprecationIssue translogRetentionSettingCheck(IndexMetadata indexMetadata, ClusterState clusterState) {
+    private DeprecationIssue translogRetentionSettingCheck(
+        IndexMetadata indexMetadata,
+        ClusterState clusterState,
+        Map<String, List<String>> ignored
+    ) {
         final boolean softDeletesEnabled = IndexSettings.INDEX_SOFT_DELETES_SETTING.get(indexMetadata.getSettings());
         if (softDeletesEnabled) {
             if (IndexSettings.INDEX_TRANSLOG_RETENTION_SIZE_SETTING.exists(indexMetadata.getSettings())
@@ -155,7 +173,7 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
         return null;
     }
 
-    private static DeprecationIssue checkIndexDataPath(IndexMetadata indexMetadata, ClusterState clusterState) {
+    private DeprecationIssue checkIndexDataPath(IndexMetadata indexMetadata, ClusterState clusterState, Map<String, List<String>> ignored) {
         if (IndexMetadata.INDEX_DATA_PATH_SETTING.exists(indexMetadata.getSettings())) {
             final String message = String.format(
                 Locale.ROOT,
@@ -170,7 +188,11 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
         return null;
     }
 
-    private static DeprecationIssue storeTypeSettingCheck(IndexMetadata indexMetadata, ClusterState clusterState) {
+    private DeprecationIssue storeTypeSettingCheck(
+        IndexMetadata indexMetadata,
+        ClusterState clusterState,
+        Map<String, List<String>> ignored
+    ) {
         final String storeType = IndexModule.INDEX_STORE_TYPE_SETTING.get(indexMetadata.getSettings());
         if (IndexModule.Type.SIMPLEFS.match(storeType)) {
             return new DeprecationIssue(
@@ -187,7 +209,11 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
         return null;
     }
 
-    private static DeprecationIssue legacyRoutingSettingCheck(IndexMetadata indexMetadata, ClusterState clusterState) {
+    private DeprecationIssue legacyRoutingSettingCheck(
+        IndexMetadata indexMetadata,
+        ClusterState clusterState,
+        Map<String, List<String>> ignored
+    ) {
         List<String> deprecatedSettings = LegacyTiersDetection.getDeprecatedFilteredAllocationSettings(indexMetadata.getSettings());
         if (deprecatedSettings.isEmpty()) {
             return null;
@@ -203,7 +229,7 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
         );
     }
 
-    private static void fieldLevelMappingIssue(IndexMetadata indexMetadata, BiConsumer<MappingMetadata, Map<String, Object>> checker) {
+    private void fieldLevelMappingIssue(IndexMetadata indexMetadata, BiConsumer<MappingMetadata, Map<String, Object>> checker) {
         if (indexMetadata.mapping() != null) {
             Map<String, Object> sourceAsMap = indexMetadata.mapping().sourceAsMap();
             checker.accept(indexMetadata.mapping(), sourceAsMap);
@@ -221,7 +247,7 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
      * @return a list of issues found in fields
      */
     @SuppressWarnings("unchecked")
-    private static List<String> findInPropertiesRecursively(
+    private List<String> findInPropertiesRecursively(
         String type,
         Map<String, Object> parentMap,
         Function<Map<?, ?>, Boolean> predicate,
@@ -275,7 +301,11 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
         return issues;
     }
 
-    private static DeprecationIssue deprecatedCamelCasePattern(IndexMetadata indexMetadata, ClusterState clusterState) {
+    private DeprecationIssue deprecatedCamelCasePattern(
+        IndexMetadata indexMetadata,
+        ClusterState clusterState,
+        Map<String, List<String>> ignored
+    ) {
         List<String> fields = new ArrayList<>();
         fieldLevelMappingIssue(
             indexMetadata,
@@ -283,8 +313,8 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
                 findInPropertiesRecursively(
                     mappingMetadata.type(),
                     sourceAsMap,
-                    IndexDeprecationChecker::isDateFieldWithCamelCasePattern,
-                    IndexDeprecationChecker::changeFormatToSnakeCase,
+                    this::isDateFieldWithCamelCasePattern,
+                    this::changeFormatToSnakeCase,
                     "",
                     ""
                 )
@@ -305,7 +335,7 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
         return null;
     }
 
-    private static boolean isDateFieldWithCamelCasePattern(Map<?, ?> property) {
+    private boolean isDateFieldWithCamelCasePattern(Map<?, ?> property) {
         if ("date".equals(property.get("type")) && property.containsKey("format")) {
             String[] patterns = DateFormatter.splitCombinedPatterns((String) property.get("format"));
             for (String pattern : patterns) {
@@ -316,7 +346,7 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
         return false;
     }
 
-    private static String changeFormatToSnakeCase(String type, Map.Entry<?, ?> entry) {
+    private String changeFormatToSnakeCase(String type, Map.Entry<?, ?> entry) {
         Map<?, ?> value = (Map<?, ?>) entry.getValue();
         final String formatFieldValue = (String) value.get("format");
         String[] patterns = DateFormatter.splitCombinedPatterns(formatFieldValue);
@@ -331,5 +361,15 @@ public class IndexDeprecationChecker implements ResourceDeprecationChecker {
         }
         sb.deleteCharAt(sb.length() - 1);
         return sb.toString();
+    }
+
+    private Map<String, List<String>> indexToTransformIds(List<TransformConfig> transformConfigs) {
+        return transformConfigs.stream()
+            .collect(
+                Collectors.groupingBy(
+                    config -> config.getDestination().getIndex(),
+                    Collectors.mapping(TransformConfig::getId, Collectors.toList())
+                )
+            );
     }
 }
