+++ x-pack/plugin/deprecation/src/main/java/org/elasticsearch/xpack/deprecation/TemplateDeprecationChecker.java
@@ -19,9 +19,9 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.function.Function;
 
-import static org.elasticsearch.xpack.deprecation.DeprecationInfoAction.filterChecks;
 import static org.elasticsearch.xpack.deprecation.LegacyTiersDetection.DEPRECATION_COMMON_DETAIL;
 import static org.elasticsearch.xpack.deprecation.LegacyTiersDetection.DEPRECATION_HELP_URL;
 import static org.elasticsearch.xpack.deprecation.LegacyTiersDetection.DEPRECATION_MESSAGE;
@@ -32,20 +32,34 @@ import static org.elasticsearch.xpack.deprecation.LegacyTiersDetection.DEPRECATI
 public class TemplateDeprecationChecker implements ResourceDeprecationChecker {
 
     public static final String NAME = "templates";
-    private static final List<Function<ComposableIndexTemplate, DeprecationIssue>> INDEX_TEMPLATE_CHECKS = List.of(
-        TemplateDeprecationChecker::checkLegacyTiersInIndexTemplate
+    private final List<Function<ComposableIndexTemplate, DeprecationIssue>> indexTemplateChecks = List.of(
+        this::checkLegacyTiersInIndexTemplate
     );
-    private static final List<Function<ComponentTemplate, DeprecationIssue>> COMPONENT_TEMPLATE_CHECKS = List.of(
-        TemplateDeprecationChecker::checkSourceModeInComponentTemplates,
-        TemplateDeprecationChecker::checkLegacyTiersInComponentTemplates
+    private final List<Function<ComponentTemplate, DeprecationIssue>> componentTemplateChecks = List.of(
+        this::checkSourceModeInComponentTemplates,
+        this::checkLegacyTiersInComponentTemplates
     );
+
+    /**
+     * @param clusterState The cluster state provided for the checker
+     * @param request not used yet in these checks
+     * @param precomputedData not used yet in these checks
+     * @return the name of the data streams that have violated the checks with their respective warnings.
+     */
+    @Override
+    public Map<String, List<DeprecationIssue>> check(
+        ClusterState clusterState,
+        DeprecationInfoAction.Request request,
+        TransportDeprecationInfoAction.PrecomputedData precomputedData
+    ) {
+        return check(clusterState);
+    }
 
     /**
      * @param clusterState The cluster state provided for the checker
      * @return the name of the data streams that have violated the checks with their respective warnings.
      */
-    @Override
-    public Map<String, List<DeprecationIssue>> check(ClusterState clusterState, DeprecationInfoAction.Request request) {
+    Map<String, List<DeprecationIssue>> check(ClusterState clusterState) {
         var indexTemplates = clusterState.metadata().templatesV2().entrySet();
         var componentTemplates = clusterState.metadata().componentTemplates().entrySet();
         if (indexTemplates.isEmpty() && componentTemplates.isEmpty()) {
@@ -56,7 +70,10 @@ public class TemplateDeprecationChecker implements ResourceDeprecationChecker {
             String name = entry.getKey();
             ComposableIndexTemplate template = entry.getValue();
 
-            List<DeprecationIssue> issuesForSingleIndexTemplate = filterChecks(INDEX_TEMPLATE_CHECKS, c -> c.apply(template));
+            List<DeprecationIssue> issuesForSingleIndexTemplate = indexTemplateChecks.stream()
+                .map(c -> c.apply(template))
+                .filter(Objects::nonNull)
+                .toList();
             if (issuesForSingleIndexTemplate.isEmpty() == false) {
                 issues.computeIfAbsent(name, ignored -> new ArrayList<>()).addAll(issuesForSingleIndexTemplate);
             }
@@ -65,7 +82,10 @@ public class TemplateDeprecationChecker implements ResourceDeprecationChecker {
             String name = entry.getKey();
             ComponentTemplate template = entry.getValue();
 
-            List<DeprecationIssue> issuesForSingleIndexTemplate = filterChecks(COMPONENT_TEMPLATE_CHECKS, c -> c.apply(template));
+            List<DeprecationIssue> issuesForSingleIndexTemplate = componentTemplateChecks.stream()
+                .map(c -> c.apply(template))
+                .filter(Objects::nonNull)
+                .toList();
             if (issuesForSingleIndexTemplate.isEmpty() == false) {
                 issues.computeIfAbsent(name, ignored -> new ArrayList<>()).addAll(issuesForSingleIndexTemplate);
             }
@@ -73,7 +93,7 @@ public class TemplateDeprecationChecker implements ResourceDeprecationChecker {
         return issues.isEmpty() ? Map.of() : issues;
     }
 
-    static DeprecationIssue checkLegacyTiersInIndexTemplate(ComposableIndexTemplate composableIndexTemplate) {
+    private DeprecationIssue checkLegacyTiersInIndexTemplate(ComposableIndexTemplate composableIndexTemplate) {
         Template template = composableIndexTemplate.template();
         if (template != null) {
             List<String> deprecatedSettings = LegacyTiersDetection.getDeprecatedFilteredAllocationSettings(template.settings());
@@ -93,7 +113,7 @@ public class TemplateDeprecationChecker implements ResourceDeprecationChecker {
         return null;
     }
 
-    static DeprecationIssue checkSourceModeInComponentTemplates(ComponentTemplate template) {
+    private DeprecationIssue checkSourceModeInComponentTemplates(ComponentTemplate template) {
         if (template.template().mappings() != null) {
             var sourceAsMap = (Map<?, ?>) XContentHelper.convertToMap(template.template().mappings().uncompressed(), true).v2().get("_doc");
             if (sourceAsMap != null) {
@@ -115,7 +135,7 @@ public class TemplateDeprecationChecker implements ResourceDeprecationChecker {
         return null;
     }
 
-    static DeprecationIssue checkLegacyTiersInComponentTemplates(ComponentTemplate componentTemplate) {
+    private DeprecationIssue checkLegacyTiersInComponentTemplates(ComponentTemplate componentTemplate) {
         Template template = componentTemplate.template();
         List<String> deprecatedSettings = LegacyTiersDetection.getDeprecatedFilteredAllocationSettings(template.settings());
         if (deprecatedSettings.isEmpty()) {
