+++ x-pack/plugin/deprecation/src/test/java/org/elasticsearch/xpack/deprecation/NodeDeprecationChecksTests.java
@@ -30,9 +30,11 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import static org.elasticsearch.xpack.deprecation.DeprecationChecks.NODE_SETTINGS_CHECKS;
+import static org.elasticsearch.xpack.deprecation.NodeDeprecationChecks.SINGLE_NODE_CHECKS;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.not;
@@ -154,8 +156,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
             .put(Environment.PATH_SHARED_DATA_SETTING.getKey(), createTempDir())
             .build();
 
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
         final String expectedUrl =
@@ -209,8 +211,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
         }
 
         final Settings settings = builder.build();
-        final List<DeprecationIssue> deprecationIssues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        final List<DeprecationIssue> deprecationIssues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
 
@@ -235,8 +237,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
 
     void monitoringSetting(String settingKey, String value) {
         Settings settings = Settings.builder().put(settingKey, value).build();
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
         final String expectedUrl = "https://ela.st/es-deprecation-7-monitoring-settings";
@@ -259,8 +261,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
         String settingKey = "xpack.monitoring.exporters.test." + suffix;
         Settings settings = Settings.builder().put(settingKey, value).build();
         final XPackLicenseState licenseState = new XPackLicenseState(() -> 0);
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, licenseState)
         );
         final String expectedUrl = "https://ela.st/es-deprecation-7-monitoring-settings";
@@ -284,8 +286,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
         String subSettingKey = settingKey + ".subsetting";
         Settings settings = Settings.builder().put(subSettingKey, value).build();
         final XPackLicenseState licenseState = new XPackLicenseState(() -> 0);
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, licenseState)
         );
         final String expectedUrl = "https://ela.st/es-deprecation-7-monitoring-settings";
@@ -310,8 +312,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
         secureSettings.setString(settingKey, value);
         Settings settings = Settings.builder().setSecureSettings(secureSettings).build();
         final XPackLicenseState licenseState = new XPackLicenseState(() -> 0);
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, licenseState)
         );
         final String expectedUrl = "https://ela.st/es-deprecation-7-monitoring-settings";
@@ -457,8 +459,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
     public void testExporterUseIngestPipelineSettings() {
         Settings settings = Settings.builder().put("xpack.monitoring.exporters.test.use_ingest", true).build();
 
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
 
@@ -483,8 +485,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
             .put("xpack.monitoring.exporters.test.index.pipeline.master_timeout", TimeValue.timeValueSeconds(10))
             .build();
 
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
 
@@ -508,8 +510,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
     public void testExporterCreateLegacyTemplateSetting() {
         Settings settings = Settings.builder().put("xpack.monitoring.exporters.test.index.template.create_legacy_templates", true).build();
 
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
 
@@ -535,8 +537,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
             .put(ScriptService.SCRIPT_GENERAL_MAX_COMPILATIONS_RATE_SETTING.getKey(), "use-context")
             .build();
 
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
 
@@ -564,8 +566,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
             .put(ScriptService.SCRIPT_MAX_COMPILATIONS_RATE_SETTING.getConcreteSettingForNamespace(contexts.get(1)).getKey(), "456/7m")
             .build();
 
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
 
@@ -601,8 +603,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
             .put(ScriptService.SCRIPT_CACHE_SIZE_SETTING.getConcreteSettingForNamespace(contexts.get(1)).getKey(), "2453")
             .build();
 
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
 
@@ -639,8 +641,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
             .put(ScriptService.SCRIPT_CACHE_SIZE_SETTING.getConcreteSettingForNamespace(contexts.get(1)).getKey(), 200)
             .build();
 
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
 
@@ -676,8 +678,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
             .put(ScriptService.SCRIPT_CACHE_EXPIRE_SETTING.getConcreteSettingForNamespace(contexts.get(1)).getKey(), "2d")
             .build();
 
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
 
@@ -708,8 +710,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
     public void testEnforceDefaultTierPreferenceSetting() {
         Settings settings = Settings.builder().put(DataTier.ENFORCE_DEFAULT_TIER_PREFERENCE_SETTING.getKey(), randomBoolean()).build();
 
-        List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            NODE_SETTINGS_CHECKS,
+        List<DeprecationIssue> issues = filterChecks(
+            SINGLE_NODE_CHECKS,
             c -> c.apply(settings, null, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
 
@@ -731,8 +733,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
     }
 
     private List<DeprecationIssue> getDeprecationIssues(Settings settings, PluginsAndModules pluginsAndModules) {
-        final List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            DeprecationChecks.NODE_SETTINGS_CHECKS,
+        final List<DeprecationIssue> issues = filterChecks(
+            NodeDeprecationChecks.SINGLE_NODE_CHECKS,
             c -> c.apply(settings, pluginsAndModules, ClusterState.EMPTY_STATE, new XPackLicenseState(() -> 0))
         );
 
@@ -799,8 +801,8 @@ public class NodeDeprecationChecksTests extends ESTestCase {
         }
         Metadata metadata = metadataBuilder.build();
         ClusterState clusterState = ClusterState.builder(ClusterName.DEFAULT).metadata(metadata).build();
-        final List<DeprecationIssue> issues = DeprecationChecks.filterChecks(
-            DeprecationChecks.NODE_SETTINGS_CHECKS,
+        final List<DeprecationIssue> issues = filterChecks(
+            NodeDeprecationChecks.SINGLE_NODE_CHECKS,
             c -> c.apply(nodettings, pluginsAndModules, clusterState, licenseState)
         );
 
@@ -831,5 +833,9 @@ public class NodeDeprecationChecksTests extends ESTestCase {
             null
         );
         assertThat(issues, hasItem(expected));
+    }
+
+    static <T> List<DeprecationIssue> filterChecks(List<T> checks, Function<T, DeprecationIssue> mapper) {
+        return checks.stream().map(mapper).filter(Objects::nonNull).toList();
     }
 }
